TouchCarControl is based on float instead of int. This allows the control to be implemented for the range 0-1.

+++ cars/src/main/java/dk/aau/cs/giraf/cars/game/Controller/TouchCarControl.java
@@ -5,9 +5,9 @@ import dk.aau.cs.giraf.cars.game.Interfaces.CarControl;
 
 
 public class TouchCarControl implements CarControl {
-    private int lastMove;
-    private final int height;
-    private final int offset;
+    private float lastMove;
+    private final float height;
+    private final float offset;
 
     public TouchCarControl(int height, int offset) {
         this.height = height;
@@ -18,9 +18,9 @@ public class TouchCarControl implements CarControl {
     public float getMove(Input.TouchEvent[] touchEvents) {
         for (Input.TouchEvent e : touchEvents)
             if (e.type == Input.TouchEvent.TOUCH_DOWN || e.type == Input.TouchEvent.TOUCH_DRAGGED)
-                lastMove = 1 - ((e.y - offset) / height);
+                lastMove = 1f - ((e.y - offset) / height);
             else if (e.type == Input.TouchEvent.TOUCH_UP)
-                lastMove = 0;
+                lastMove = 0f;
 
         return lastMove;
     }
